+++ src/example/capstone/datatypes/UpdateOperation.java
@@ -7,9 +7,9 @@ import java.util.Map;
 
 public class UpdateOperation extends Operation {
 
-    private Map<Integer, Integer> clientClock;
+    private Map<Long, Integer> clientClock;
 
-    public UpdateOperation(int key, Map<Integer, Integer> clientClock) {
+    public UpdateOperation(int key, Map<Long, Integer> clientClock) {
         super(Type.UPDATE, key);
         this.clientClock = new HashMap<>(clientClock);
     }
